refactor: 값 객체 한곳에 모으기
