+++ src/main/java/domain/value/BonusNumber.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.value;
+
+import domain.Lotto;
 
 import java.util.List;
 public record BonusNumber(int value) {
