+++ src/main/java/domain/value/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.value;
 public record Money(int value) {
     public Money {
         if (value <= 0) throw new IllegalArgumentException("구매 금액은 양수여야 합니다.");
